requirements.txt
moviepy
instagrapi
schedule
python-dotenv